Added custom rule group interval
* feat: adjustable rule_group interval

This allows for more expensive SLO recording rules by letting users
create a custom Prometheus `rule_group.interval`.

https://prometheus.io/docs/prometheus/latest/configuration/recording_rules/#rule_group

Now instead of all rule groups using the default global evaluation
interval, a custom interval can be set on all sets of recording rules
for an SLO.

If no `interval` is set then the global default will be assumed,
matching current Sloth behavior.

Resolves: https://github.com/slok/sloth/issues/367

* docs: add example file for custom rule groups

* ref: only append RuleGroupInterval when not empty

Related to how the unit test coverage works, this will make sure
RuleGroupInterval is not appended to RuleGroups unless it is defined and
not empty.

Not requring the `yaml` file does an okay job of this but expliclity
writing the rules without RuleGroupInterval is safer.

It's a bit ugly and repetitive but I think that's just how Go works..

* test: update tests for custom rule_group interval

This updates tests include `interval`'s and also includes a test to make
sure the rules render correctly when interval is not included.

* ref: add rule intervals per rule type

Instead of a singular global default, now a rule_group interval can be
set for every individual type of rule_group Sloth generates.

The generic, `interval:all` rule will also stay and can "fill in" any missing per-rule
group defaults. Along with the default behavior of doing nothing if no
`interval` is specified.

* test: add tests for custom rule group intervals

Specific ones and general/specific combined

* chore: update example for custom rule groups

* docs: finish comments

Co-authored-by: Will Hegedus <will@wbhegedus.me>

* ref: rule interval groups less repetitive

* chore: golint else outindent

---------

Co-authored-by: Will Bollock <wbollock@linode.com>
Co-authored-by: Will Bollock <wbollock@gmail.com>
Co-authored-by: Will Hegedus <will@wbhegedus.me>
Signed-off-by: Синявский Дмитрий Сергеевич <dmitriy.sinyavskiy@vitech.team>